+++ js/battle/character/index.js
@@ -1,0 +1,15 @@
+export { Character } from "./character.js";
+export { CharacterController } from "./character-controller.js";
+export { CharacterInput } from "./character-input.js";
+export { CharacterSystem, createCharacterSystem } from "./character-system.js";
+export { CharacterView, RemoteCharacterView, createCharacterActorElement } from "./character-view.js";
+export { VirtualJoystick } from "./virtual-joystick.js";
+export {
+  CHARACTER_CONTACT_PHASES,
+  CHARACTER_DIRECTIONS,
+  CHARACTER_DIRECTION_VALUES,
+  CHARACTER_EVENTS,
+  CHARACTER_STATES,
+  CHARACTER_STATE_VALUES,
+  CHARACTER_TRIGGER_KINDS,
+} from "./constants.js";